+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - `View.escape` to escape the tracker zone for `EventHandler`s that are registered inside the `View`.
 - Expose `View` in `BuildContext`.
 - Experimental `SubView`.
-- Experimental `StatefulComponent`.
+- Experimental `StatefulComponent` (moved to `experimental.dart`).
 
 ## 0.2.0+1